fix(): Separate packages according to class attribute
