+++ src/main/java/step1/racing/common/RandomIntegerGenerator.java
@@ -1,4 +1,4 @@
-package step1.racing;
+package step1.racing.common;
 
 import java.util.Random;
