+++ src/main/java/step1/racing/common/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package step1.racing;
+package step1.racing.common;
 
 public interface RandomNumberGenerator<T extends Number> {
     T gacha(int bound);
