+++ src/main/java/step1/racing/data/RacingCar.java
@@ -1,7 +1,5 @@
 package step1.racing.data;
 
-import java.util.stream.IntStream;
-
 public class RacingCar {
     private final String name;
     private int position;
@@ -30,13 +28,5 @@ public class RacingCar {
 
     public void moveForward() {
         position++;
-    }
-
-    public String printStatus(String flag) {
-        String status = IntStream.range(0, position)
-                                 .mapToObj(pos -> flag)
-                                 .reduce(String::concat).get();
-
-        return name + " : " + status;
     }
 }
